requirements.txt
ollama>=0.6.0,<1.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
